Menu OK, guard ok

+++ src/app/views/admin/product-list/product-list.component.ts
@@ -6,18 +6,22 @@ import { Subscription } from 'rxjs';
 import { BreadcrumbsComponent } from 'src/app/shared/component/breadcrumbs/breadcrumbs.component';
 import { Product } from 'src/app/api/models/class/product';
 import { ProductService } from 'src/app/api/services/product.service';
+import { ButtonModule } from 'primeng/button';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
   standalone: true,
-  imports: [CommonModule, FormsModule, ReactiveFormsModule, BreadcrumbsComponent]
+  imports: [CommonModule, FormsModule, ReactiveFormsModule, BreadcrumbsComponent, ButtonModule]
 })
 export class ProductListComponent implements OnInit, OnDestroy {
   products!: any[];
   productSubscription!: Subscription;
 
   constructor(
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService,
     private productService: ProductService,
     private router: Router
   ) {}
@@ -29,6 +33,11 @@ export class ProductListComponent implements OnInit, OnDestroy {
     this.productService.emitProducts();
   }
 
+  ellipsis(text: string, maxLength: number): string {
+    if (!text) return '';
+    return text.length > maxLength ? text.substring(0, maxLength) + '...' : text;
+  }
+
   onNewProduct() {
     this.router.navigate(['/products', 'new']);
   }
@@ -37,8 +46,39 @@ export class ProductListComponent implements OnInit, OnDestroy {
     this.router.navigate(['/products', 'edit', key]);
   }
 
-  onDeleteProduct(product: Product) {
-    this.productService.removeProduct(product);
+  onDeleteProduct(event: Event, product: Product) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Voulez vous vraiment supprimer le produit: ' + product.title,
+      header: 'Attention!',
+      icon: 'pi pi-exclamation-triangle',
+      rejectLabel: 'Annuler',
+      rejectButtonProps: {
+        label: 'Annuler',
+        severity: 'secondary',
+        outlined: true
+      },
+      acceptButtonProps: {
+        label: 'Supprimer',
+        severity: 'danger'
+      },
+
+      accept: () => {
+        this.productService.removeProduct(product);
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Confirmation',
+          detail: 'Le produit ' + product.title + 'a bien été supprimé'
+        });
+      },
+      reject: () => {
+        this.messageService.add({
+          severity: 'severity',
+          summary: 'Annulé',
+          detail: 'La demande a bien été annulé'
+        });
+      }
+    });
   }
 
   onViewProduct(key: string) {
